issue: move plugin depdending podvolume functions to util pkg
